+++ src/base/parser/grammar/ast_printer.h
@@ -40,7 +40,7 @@ public:
 
     void operator() (std::string const&) const;
     void operator() (VarId    const&) const;
-    void operator() (int      const&) const;
+    void operator() (int64_t  const&) const;
     void operator() (float    const&) const;
     void operator() (bool     const&) const;
     void operator() (ast::And const&) const;
